+++ cognite/client/_constants.py
@@ -1,0 +1,5 @@
+# Functions
+LIST_LIMIT_DEFAULT = 25
+LIST_LIMIT_CEILING = 10_000  # variable used to guarantee all items are returned when list(limit) is None, inf or -1.
+HANDLER_FILE_NAME = "handler.py"
+MAX_RETRIES = 5